requirements.txt
torch
torchvision
tqdm
Pillow
torchsummary